requirements.txt
python-dotenv==1.0.0
pandas==2.0.3
twilio==8.4.0
tabulate==0.9.0